Add comment field to festival model & ownership
Add `comments` field to `festival`'s model so that the
backend will be expecting a comment. Add `comment` to the user model
so that the backend will be aware that users own comments. Create
`festivalSchema` variable in user model so that the user's schema
knows that users can have data ownership.

Next steps: Ensure that delete-related functions are dependent. (i.e
make sure that a comment is deleted when a festival is deleted.)